feat(core): Update workflow review request schema and related entities
- Changed 'status' to 'state' with values 'open' and 'closed'.
- Introduced 'decision' field with values 'pending', 'changes_requested', and 'approved'.
- Replaced 'archivedAt' and 'archivedById' with 'closedAt' and 'closedById'.
- Updated related entities and migration scripts to reflect schema changes.
- Adjusted repository methods and tests to accommodate new fields and logic.
