+++ packages/@n8n/db/src/entities/index.ts
@@ -66,8 +66,10 @@ import { WorkflowReviewRequestReviewer } from './workflow-review-request-reviewe
 import { WorkflowReviewRequestWorkflow } from './workflow-review-request-workflow.ee';
 import {
 	WorkflowReviewRequest,
-	WorkflowReviewRequestStatus,
-	WorkflowReviewRequestStatusList,
+	WorkflowReviewRequestDecision,
+	WorkflowReviewRequestDecisionList,
+	WorkflowReviewRequestState,
+	WorkflowReviewRequestStateList,
 } from './workflow-review-request.ee';
 import { WorkflowStatistics } from './workflow-statistics';
 import { WorkflowTagMapping } from './workflow-tag-mapping';
@@ -124,8 +126,10 @@ export {
 	WorkflowPublishedVersion,
 	WorkflowPublishHistory,
 	WorkflowReviewRequest,
-	WorkflowReviewRequestStatus,
-	WorkflowReviewRequestStatusList,
+	WorkflowReviewRequestState,
+	WorkflowReviewRequestStateList,
+	WorkflowReviewRequestDecision,
+	WorkflowReviewRequestDecisionList,
 	WorkflowReviewRequestWorkflow,
 	WorkflowReviewRequestReviewer,
 	WorkflowReviewRequestAuthor,
