+++ packages/@n8n/db/src/entities/workflow-review-request-author.ee.ts
@@ -1,12 +1,12 @@
-import { Column, Entity, Index } from '@n8n/typeorm';
+import { Column, Entity, Unique } from '@n8n/typeorm';
 
 import { WithStringId } from './abstract-entity';
 
 @Entity({ name: 'workflow_review_request_authors' })
-@Index('IDX_workflow_review_request_authors_request_user', ['workflowReviewRequestId', 'userId'], {
-	unique: true,
-})
-@Index('IDX_workflow_review_request_authors_user_id', ['userId'])
+@Unique('UQ_workflow_review_request_authors_request_user', [
+	'workflowReviewRequestId',
+	'userId',
+])
 export class WorkflowReviewRequestAuthor extends WithStringId {
 	@Column({ type: 'varchar', length: 36 })
 	workflowReviewRequestId: string;
