+++ packages/@n8n/db/src/entities/workflow-review-request-reviewer.ee.ts
@@ -1,16 +1,12 @@
-import { Column, Entity, Index } from '@n8n/typeorm';
+import { Column, Entity, Unique } from '@n8n/typeorm';
 
 import { WithStringId } from './abstract-entity';
 
 @Entity({ name: 'workflow_review_request_reviewers' })
-@Index(
-	'IDX_workflow_review_request_reviewers_request_user',
-	['workflowReviewRequestId', 'userId'],
-	{
-		unique: true,
-	},
-)
-@Index('IDX_workflow_review_request_reviewers_user_id', ['userId'])
+@Unique('UQ_workflow_review_request_reviewers_request_user', [
+	'workflowReviewRequestId',
+	'userId',
+])
 export class WorkflowReviewRequestReviewer extends WithStringId {
 	@Column({ type: 'varchar', length: 36 })
 	workflowReviewRequestId: string;
